Revert "WarmVmCache"
This reverts commit 3b99bb75aa68c48814559aafbc3a351c64b45b37.
