+++ aptos-move/aptos-vm/src/lib.rs
@@ -120,6 +120,7 @@ pub mod testing;
 pub mod transaction_metadata;
 mod transaction_validation;
 mod verifier;
+mod warm_vm_cache;
 
 pub use crate::aptos_vm::AptosVM;
 use crate::sharded_block_executor::{executor_client::ExecutorClient, ShardedBlockExecutor};
