+++ aptos-move/aptos-vm/src/move_vm_ext/mod.rs
@@ -7,7 +7,6 @@ mod resolver;
 mod respawned_session;
 mod session;
 mod vm;
-mod warm_vm_cache;
 pub(crate) mod write_op_converter;
 
 pub use crate::move_vm_ext::{
